Move error logging impls to own module
